+++ gram/OTMulti.h
@@ -25,7 +25,7 @@
 
 #include "OTMulti_def.h"
 
-long OTMulti_getConstraintIndexFromName (OTMulti me, const char32 *name);
+integer OTMulti_getConstraintIndexFromName (OTMulti me, const char32 *name);
 
 void OTMulti_checkIndex (OTMulti me);
 
@@ -38,9 +38,9 @@ void OTMulti_sort (OTMulti me);
 
 void OTMulti_newDisharmonies (OTMulti me, double evaluationNoise);
 
-int OTMulti_candidateMatches (OTMulti me, long icand, const char32 *form1, const char32 *form2);
-int OTMulti_compareCandidates (OTMulti me, long icand1, long icand2);
-long OTMulti_getWinner (OTMulti me, const char32 *form1, const char32 *form2);
+int OTMulti_candidateMatches (OTMulti me, integer icand, const char32 *form1, const char32 *form2);
+int OTMulti_compareCandidates (OTMulti me, integer icand1, integer icand2);
+integer OTMulti_getWinner (OTMulti me, const char32 *form1, const char32 *form2);
 
 #define OTMulti_LEARN_FORWARD  1
 #define OTMulti_LEARN_BACKWARD  2
@@ -49,8 +49,8 @@ int OTMulti_learnOne (OTMulti me, const char32 *form1, const char32 *form2,
 	enum kOTGrammar_rerankingStrategy updateRule, int direction, double plasticity, double relativePlasticityNoise);
 void OTMulti_PairDistribution_learn (OTMulti me, PairDistribution thee,
 	double evaluationNoise, enum kOTGrammar_rerankingStrategy updateRule, int direction,
-	double initialPlasticity, long replicationsPerPlasticity, double plasticityDecrement,
-	long numberOfPlasticities, double relativePlasticityNoise, long storeHistoryEvery, autoTable *history_out);
+	double initialPlasticity, integer replicationsPerPlasticity, double plasticityDecrement,
+	integer numberOfPlasticities, double relativePlasticityNoise, integer storeHistoryEvery, autoTable *history_out);
 
 void OTMulti_drawTableau (OTMulti me, Graphics g, const char32 *form1, const char32 *form2, bool vertical, bool showDisharmonies);
 
@@ -61,13 +61,13 @@ autoOTMulti OTMulti_create_metrics (
 	int includeClashAndLapse, int includeCodas);
 
 void OTMulti_reset (OTMulti me, double ranking);
-void OTMulti_setRanking (OTMulti me, long constraint, double ranking, double disharmony);
-void OTMulti_setConstraintPlasticity (OTMulti me, long constraint, double plasticity);
+void OTMulti_setRanking (OTMulti me, integer constraint, double ranking, double disharmony);
+void OTMulti_setConstraintPlasticity (OTMulti me, integer constraint, double plasticity);
 void OTMulti_removeConstraint (OTMulti me, const char32 *constraintName);
 
 void OTMulti_generateOptimalForm (OTMulti me, const char32 *form1, const char32 *form2, char32 *optimalForm, double evaluationNoise);
-autoStrings OTMulti_generateOptimalForms (OTMulti me, const char32 *form1, const char32 *form2, long numberOfTrials, double evaluationNoise);
-autoDistributions OTMulti_to_Distribution (OTMulti me, const char32 *form1, const char32 *form2, long numberOfTrials, double evaluationNoise);
+autoStrings OTMulti_generateOptimalForms (OTMulti me, const char32 *form1, const char32 *form2, integer numberOfTrials, double evaluationNoise);
+autoDistributions OTMulti_to_Distribution (OTMulti me, const char32 *form1, const char32 *form2, integer numberOfTrials, double evaluationNoise);
 autoStrings OTMulti_Strings_generateOptimalForms (OTMulti me, Strings forms, double evaluationNoise);
 
 /* End of file OTMulti.h */
