+++ stat/Distributions_and_Strings.h
@@ -1,6 +1,6 @@
 /* Distributions_and_Strings.h
  *
- * Copyright (C) 1997-2011,2015 Paul Boersma
+ * Copyright (C) 1997-2011,2015,2017 Paul Boersma
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -19,8 +19,8 @@
 #include "Distributions.h"
 #include "Strings_.h"
 
-autoStrings Distributions_to_Strings (Distributions me, long column, long numberOfStrings);
-autoStrings Distributions_to_Strings_exact (Distributions me, long column);
+autoStrings Distributions_to_Strings (Distributions me, integer column, integer numberOfStrings);
+autoStrings Distributions_to_Strings_exact (Distributions me, integer column);
 autoDistributions Strings_to_Distributions (Strings me);
 
 /* End of file Distributions_and_Strings.h */
